Added compatibility for Lootr.
Lootr provides instanced inventories for all chests & barrels that have
an associated loot table. Each time a player opens a chest for the first
time, a new inventory is generated & populated.

This results in less stress on multiplayer servers as people are less
likely to explore long distances because "all the structures have been
looted".

The static LockableLootTileEntity.setLootTable is a hook that has the correct
worldgen reader/writer passed in to it, whereas unfortunately it's
impossible to hook off the non-static method.

Otherwise, the functionality is identical, and, if anything, duplicated
by not using the static method.